split types part 2
https://coveord.atlassian.net/browse/KIT-2600
