+++ packages/headless/src/app/ssr-engine/common-ssr-types.ts
@@ -1,4 +1,15 @@
 import {Controller} from '../../controllers';
+import {CoreEngine} from '../engine';
+
+export type HasKeys<TObject> = TObject extends {}
+  ? keyof TObject extends never
+    ? false
+    : true
+  : boolean;
+
+export interface OptionsExtender<TOptions> {
+  (options: TOptions): TOptions | Promise<TOptions>;
+}
 
 export interface ControllersPropsMap {
   [customName: string]: unknown;
@@ -8,6 +19,82 @@ export interface ControllersMap {
   [customName: string]: Controller;
 }
 
-export interface OptionsExtender<TOptions> {
-  (options: TOptions): TOptions | Promise<TOptions>;
+export interface ControllerSnapshot<TState> {
+  initialState: TState;
 }
+
+export interface ControllerSnapshotsMap {
+  [customName: string]: ControllerSnapshot<unknown>;
+}
+export interface ControllerDefinitionWithoutProps<
+  TEngine extends CoreEngine,
+  TController extends Controller
+> {
+  build(engine: TEngine): TController;
+}
+
+export interface ControllerDefinitionWithProps<
+  TEngine extends CoreEngine,
+  TController extends Controller,
+  TProps
+> {
+  buildWithProps(engine: TEngine, props: TProps): TController;
+}
+
+export type ControllerDefinition<
+  TEngine extends CoreEngine,
+  TController extends Controller
+> =
+  | ControllerDefinitionWithoutProps<TEngine, TController>
+  | ControllerDefinitionWithProps<TEngine, TController, unknown>;
+
+export interface ControllerDefinitionsMap<
+  TEngine extends CoreEngine,
+  TController extends Controller
+> {
+  [customName: string]: ControllerDefinition<TEngine, TController>;
+}
+
+export interface EngineAndControllers<
+  TEngine extends CoreEngine,
+  TControllers extends ControllersMap
+> {
+  engine: TEngine;
+  controllers: TControllers;
+}
+
+export type InferControllerPropsFromDefinition<
+  TController extends ControllerDefinition<CoreEngine, Controller>
+> = TController extends ControllerDefinitionWithProps<
+  CoreEngine,
+  Controller,
+  infer Props
+>
+  ? Props
+  : TController extends ControllerDefinitionWithoutProps<CoreEngine, Controller>
+  ? {}
+  : unknown;
+
+export type InferControllerPropsMapFromDefinitions<
+  TControllers extends ControllerDefinitionsMap<CoreEngine, Controller>
+> = {
+  [K in keyof TControllers as HasKeys<
+    InferControllerPropsFromDefinition<TControllers[K]>
+  > extends false
+    ? never
+    : K]: InferControllerPropsFromDefinition<TControllers[K]>;
+};
+
+export type InferControllerFromDefinition<
+  TDefinition extends ControllerDefinition<CoreEngine, Controller>
+> = TDefinition extends ControllerDefinition<infer _, infer TController>
+  ? TController
+  : never;
+
+export type InferControllersMapFromDefinition<
+  TControllers extends ControllerDefinitionsMap<CoreEngine, Controller>
+> = {[K in keyof TControllers]: InferControllerFromDefinition<TControllers[K]>};
+
+export type InferControllerSnapshotsMapFromDefinitions<
+  TControllers extends ControllerDefinitionsMap<CoreEngine, Controller>
+> = {[K in keyof TControllers]: {initialState: TControllers[K]}};
