+++ packages/headless/src/ssr.index.ts
@@ -5,6 +5,6 @@ polyfillCryptoNode();
 // 3rd Party Libraries
 export type {Unsubscribe, Middleware} from '@reduxjs/toolkit';
 
-export type {InferHydrationResult} from './app/ssr-engine/ssr-engine';
+export type {InferHydrationResult} from './app/ssr-engine/hydrate-initial-state-ssr-types';
 
 export {defineSearchEngine} from './app/ssr-engine/ssr-engine';
